# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Docseq DMS API
  description: |
    Sistema de Gestão de Documentos Municipais.
    
    **Base URL:** `http://localhost:8080`
  version: 1.0.0
  contact:
    name: Docseq DMS
servers:
  - url: http://localhost:8080/api/v1
    description: Desenvolvimento local
...
paths:
...
  /users/{id}/permissions:
    get:
      summary: Obter permissões do usuário
      operationId: getUserPermissions
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Permissões do usuário
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserPermission'
    put:
      summary: Atualizar permissões do usuário
      operationId: updateUserPermissions
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserPermissionsInput'
      responses:
        '200':
          description: Permissões atualizadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserPermission'

  # ─────────────────────────────────────────────
  # POST /auth/login
  # ─────────────────────────────────────────────
...
  /users:
    post:
      summary: Criar usuário
      operationId: createUser
      tags: [users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserInput'
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithPasswordResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      summary: Listar usuários
      operationId: listUsers
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: search
          in: query
          required: false
          description: Filtro de busca textual por username ou e-mail
          schema:
            type: string
        - name: role
          in: query
          required: false
          description: Filtro por perfil do usuário (ADMIN, MOD, COMMON)
          schema:
            type: string
            enum: [ADMIN, MOD, COMMON]
        - name: municipalityId
          in: query
          required: false
          description: Filtro por ID do município
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista paginada de usuários
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsersResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/me/change-password:
    post:
      summary: Alterar senha do usuário autenticado
      operationId: changePassword
      tags: [users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordInput'
      responses:
        '200':
          description: Senha alterada com sucesso. Retorna novo token JWT e dados do usuário.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Não autenticado ou senha atual incorreta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/trash:
    get:
      summary: Listar usuários excluídos (lixeira)
      operationId: listDeletedUsers
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: search
          in: query
          required: false
          description: Filtro de busca textual por username ou e-mail
          schema:
            type: string
        - name: role
          in: query
          required: false
          description: Filtro por perfil do usuário (ADMIN, MOD, COMMON)
          schema:
            type: string
            enum: [ADMIN, MOD, COMMON]
        - name: municipalityId
          in: query
          required: false
          description: Filtro por ID do município
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista paginada de usuários na lixeira
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsersResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/{id}:
    get:
      summary: Buscar usuário por ID
      operationId: getUserByID
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Usuário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      summary: Atualizar usuário parcialmente
      operationId: updateUser
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserInput'
      responses:
        '200':
          description: Usuário atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Excluir usuário (soft delete)
      operationId: deleteUser
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Usuário excluído com sucesso
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/{id}/restore:
    patch:
      summary: Restaurar usuário da lixeira
      operationId: restoreUser
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Usuário restaurado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/{id}/hard:
    delete:
      summary: Excluir usuário permanentemente
      operationId: hardDeleteUser
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Usuário excluído permanentemente
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ─────────────────────────────────────────────
  # POST/GET /documents
  # ─────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Parameters ─────────────────────────────
  parameters:
...
    Page:
      name: page
      in: query
      required: false
      description: Número da página (começa em 1)
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      description: Itens por página (máx. 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20

  # ─── Schemas ────────────────────────────────
  schemas:
...
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            token:
              type: string
            user:
              $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [ADMIN, MOD, COMMON]
        mustChangePassword:
          type: boolean
        municipalityId:
          type: string
          format: uuid
        lastLogin:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/User'
    UserWithPasswordResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            randomPassword:
              type: string
              example: "aB3@c1D2e3F4"
    PaginatedUsersResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    CreateUserInput:
      type: object
      required: [username, email, role, municipalityId]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 255
          example: "john_doe"
        email:
          type: string
          format: email
          maxLength: 255
          example: "john@example.com"
        role:
          type: string
          enum: [ADMIN, MOD, COMMON]
          example: "COMMON"
        municipalityId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
    UpdateUserInput:
      type: object
      minProperties: 1
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 255
          example: "john_doe"
        email:
          type: string
          format: email
          maxLength: 255
          example: "john@example.com"
        password:
          type: string
          minLength: 6
          maxLength: 255
          example: "new_secret123"
        role:
          type: string
          enum: [ADMIN, MOD, COMMON]
          example: "COMMON"
        municipalityId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        resetPassword:
          type: boolean
          example: true
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword, confirmPassword]
      properties:
        currentPassword:
          type: string
          minLength: 6
          maxLength: 255
          example: "aB3@c1D2e3F4"
        newPassword:
          type: string
          minLength: 6
          maxLength: 255
          example: "newPassword123"
        confirmPassword:
          type: string
          minLength: 6
          maxLength: 255
          example: "newPassword123"
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        total:
          type: integer
          format: int64
          example: 42
        totalPages:
          type: integer
          example: 3
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "descrição do erro"
    UserPermission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        documentType:
          type: string
          enum: [NOTICE, DECREE, ORDINANCE, LAW, CONTRACT]
        level:
          type: string
          enum: [NONE, READ, WRITE, DELETE]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateUserPermissionsInput:
      type: object
      required: [permissions]
      properties:
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/UpdateUserPermissionItem'
    UpdateUserPermissionItem:
      type: object
      required: [documentType, level]
      properties:
        documentType:
          type: string
          enum: [NOTICE, DECREE, ORDINANCE, LAW, CONTRACT]
        level:
          type: string
          enum: [NONE, READ, WRITE, DELETE]

  # ─── Responses reutilizáveis ─────────────────
  responses:
    BadRequest:
      description: Dados inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid_uuid:
              summary: UUID malformado
              value:
                success: false
                error: "ID inválido"
            invalid_uf:
              summary: UF inexistente
              value:
                success: false
                error: "UF inválida"
            validation:
              summary: Falha de validação
              value:
                success: false
                error: "Key: 'CreateMunicipalityInput.Name' Error:Field validation for 'Name' failed on the 'required' tag"
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "município não encontrado"
    Conflict:
      description: Conflito — nome já cadastrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "já existe um município com este nome"
    InternalError:
      description: Erro interno no servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "erro interno no servidor"